LotteryRoundNumber moved to app.js
