+++ src/App.css
@@ -1,3 +1,7 @@
+body{
+  text-align: center;
+}
+
 .App {
   text-align: center;
 }
