+++ src/App.js
@@ -7,10 +7,24 @@ import Number from './components/Number/Number';
 import Navigation from './components/Navigation';
 import LoginRedirect from './components/Login/LoginRedirect';
 
+import request from "./request";
+
 import LoginState from './components/Login/LoginState';
 import MyNumber from './components/Number/MyNumber';
+import LotteryRoundNumber from './components/Number/LotteryRoundNumber';
 
 function App() {
+
+  const [lotteryRound, setLotteryRound] = useState('');
+
+  useEffect(() =>{
+    const getLotteryRound = async () =>{
+        const lotteryRoundData = await request.getLotteryRound();
+        setLotteryRound(lotteryRoundData);
+    }
+
+    getLotteryRound();
+  }, [])
 
   return (
     <div>
@@ -19,15 +33,17 @@ function App() {
         <LoginState/>
         <Navigation/>
       </div>
-      
+      <div>
+        <LotteryRoundNumber/>
+      </div>
 
       <Routes>
         <Route path="/" element={<Home/>}/>
-        <Route path="/number" element={<Number/>}/>
+        <Route path="/number" element={<Number currentRound={lotteryRound}/>}/>
         <Route path="/my/number" element={<MyNumber/>}/>
         <Route path="/login/code/kakao" element={<LoginRedirect/>}/>
       </Routes>
-
+      
     </div>
     
   );
